Clean up layout and fix loading/error/success states
